restrict users to delete and update someone's announcement

+++ src/main/java/az/mapacademy/announcement/service/AnnouncementService.java
@@ -12,6 +12,7 @@ import az.mapacademy.announcement.mapper.AnnouncementMapper;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.data.domain.Page;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
@@ -52,11 +53,27 @@ public class AnnouncementService {
         return baseResponse;
     }
 
+    public BaseResponse<List<AnnouncementResponse>> getMyAllAnnouncements(
+            int page, int size) {
+        String username = getUsername();
+
+        Page<Announcement> announcementsPage = announcementDao.findAllByUsername(page, size, username);
+        List<Announcement> announcements = announcementsPage.getContent();
+        log.info("Announcements for user {}, found: {}", username, announcements);
+
+        var announcementList = announcementMapper.toResponseList(announcements);
+
+        BaseResponse<List<AnnouncementResponse>> baseResponse = new BaseResponse<>();
+        baseResponse.setData(announcementList);
+        baseResponse.setPageCount(announcementsPage.getTotalPages());
+        return baseResponse;
+    }
+
     public AnnouncementResponse createAnnouncement(CreateAnnouncementRequest request) {
         Announcement announcement = announcementMapper.toEntity(request);
         log.info("Announcement create entity: {}", announcement);
 
-        String username = SecurityContextHolder.getContext().getAuthentication().getPrincipal().toString();
+        String username = getUsername();
 
         var user = userService.getByUsername(username)
                 .orElseThrow(() -> new NotFoundException("User not found"));
@@ -70,6 +87,11 @@ public class AnnouncementService {
         Optional<Announcement> optAnnouncement = announcementDao.findById(announcementId);
         Announcement announcement = optAnnouncement.orElseThrow(() ->
                 new NotFoundException("Announcement is not found with id: " + announcementId));
+        var username = getUsername();
+
+        if (!announcement.getUser().getUsername().equals(username)) {
+            throw new AccessDeniedException("User is not allowed to update announcement");
+        }
 
         announcementMapper.populate(request, announcement);
         log.info("Announcement update entity: {}", announcement);
@@ -78,6 +100,15 @@ public class AnnouncementService {
     }
 
     public void deleteAnnouncement(Long announcementId) {
+        Optional<Announcement> optAnnouncement = announcementDao.findById(announcementId);
+        Announcement announcement = optAnnouncement.orElseThrow(() ->
+                new NotFoundException("Announcement is not found with id: " + announcementId));
+        var username = getUsername();
+
+        if (!announcement.getUser().getUsername().equals(username)) {
+            throw new AccessDeniedException("User is not allowed to delete announcement");
+        }
+
         announcementDao.delete(announcementId);
     }
 
@@ -89,5 +120,9 @@ public class AnnouncementService {
         log.info("Announcement found: {}", announcement);
 
         return announcementMapper.toResponse(announcement);
+    }
+
+    private static String getUsername() {
+        return SecurityContextHolder.getContext().getAuthentication().getPrincipal().toString();
     }
 }
